# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Errorlog:
    properties:
      error:
        format: string
        type: string
      refresh_time:
        format: time
        type: string
    type: object
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      err_code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
...
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMPCSControlRequest:
    properties:
      action:
        description: 'action: setup/config/create/enable/disable/move/cleanup/status/remove/destroy/stop/sync/ccvm-status'
        example: status
        type: string
      cluster:
        description: cluster name for config action
        example: cloudcenter_cluster
        type: string
      hosts:
        description: hostnames for config action
        example:
        - ablecube31-1
        - ablecube31-2
        - ablecube31-3
        items:
          type: string
        type: array
      resource:
        description: PCS resource name. 비어 있으면 cloudcenter_res를 사용한다.
        example: cloudcenter_res
        type: string
      target:
        description: target hostname/IP for move action
        example: ablecube31-2
        type: string
      time:
        description: cluster token time for sync action
        example: "3000"
        type: string
      xml:
        description: CCVM libvirt XML path for create action
        example: /etc/libvirt/qemu/ccvm.xml
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.CCVMPCSControlResponse:
    properties:
      action:
        example: status
        type: string
      code:
        example: 200
        type: integer
      message:
        type: string
      retname:
        type: string
      target:
        example: 10.10.31.1
        type: string
      val: {}
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMServiceControlRequest:
    properties:
      action:
        description: 'action: start/restart/stop/status'
        example: start
        type: string
      service_name:
        description: service name
        example: mold
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.CCVMServiceControlResponse:
    properties:
      code:
        example: 200
        type: integer
      val:
        example: mold service start control success
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_mold.BootstrapRequest:
    properties:
      auth:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.MoldAuthRequest'
      cluster:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.ClusterRequest'
      physical_network:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.PhysicalNetworkRequest'
      pod:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.PodRequest'
      public_ip:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.PublicIPRangeRequest'
      secondary_staging_store:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.SecondaryStagingStoreRequest'
      zone:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.ZoneRequest'
    type: object
...
  ablecloud_io_ablestack-api_internal_model_mold.ClusterRequest:
    properties:
      arch:
        example: x86_64
        type: string
      clustertype:
        example: CloudManaged
        type: string
      hypervisor:
        example: KVM
        type: string
      name:
        example: Cluster
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_mold.MoldAuthRequest:
    properties:
      domain:
        example: /
        type: string
      password:
        example: password
        type: string
      response:
        example: json
        type: string
      username:
        example: admin
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_mold.PhysicalNetworkRequest:
    properties:
      broadcastdomainrange:
        example: Zone
        type: string
      isolationmethods:
        example: VLAN
        type: string
      name:
        example: Physicalnetwork
        type: string
      networkspeed:
        example: 10G
        type: string
      tags:
        example: guest
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_mold.PodRequest:
    properties:
      endip:
        example: 10.10.31.150
        type: string
      gateway:
        example: 10.10.31.1
        type: string
      name:
        example: Pod
        type: string
      netmask:
        example: 255.255.255.0
        type: string
      startip:
        example: 10.10.31.100
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_mold.PublicIPRangeRequest:
    properties:
      endip:
        example: 10.10.20.150
        type: string
      gateway:
        example: 10.10.20.1
        type: string
      netmask:
        example: 255.255.255.0
        type: string
      startip:
        example: 10.10.20.100
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_mold.Response:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: ok
        type: string
      val: {}
    type: object
  ablecloud_io_ablestack-api_internal_model_mold.SecondaryStagingStoreRequest:
    properties:
      provider:
        example: NFS
        type: string
      scope:
        example: zone
        type: string
      url:
        example: nfs://10.10.31.20/export/staging
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_mold.ZoneRequest:
    properties:
      dns1:
        example: 8.8.8.8
        type: string
      dns2:
        example: 1.1.1.1
        type: string
      domain:
        example: local
        type: string
      guestcidraddress:
        example: 10.1.1.0/24
        type: string
      internaldns1:
        example: 10.10.0.1
        type: string
      internaldns2:
        example: 10.10.0.2
        type: string
      localstorageenabled:
        example: "false"
        type: string
      name:
        example: Zone
        type: string
      networktype:
        example: Advanced
        type: string
      securitygroupenabled:
        example: "false"
        type: string
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/ccvm/service/control:
    post:
      consumes:
      - application/json
      description: CCVM의 서비스 제어 요청을 CCVM 노드로 전달합니다.
      parameters:
      - description: ccvm service control request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMServiceControlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMServiceControlResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: CCVM Service Control
      tags:
      - Cube-CCVM
...
  /cube/pcs/control:
    post:
      consumes:
      - application/json
      description: CCVM Pacemaker/PCS setup/config/create/enable/disable/move/cleanup/status/remove/destroy/stop/sync/ccvm-status
        작업을 수행합니다. setup은 clusterConfig.pcsCluster 노드의 기존 Python 스냅샷 cron을 정리한 뒤 cloudcenter_res를
        구성합니다. 자동 스냅샷은 API 내부 Go 스케줄러가 수행합니다.
      parameters:
      - description: pcs control request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMPCSControlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMPCSControlResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: PCS Control
      tags:
      - Cube-PCS
...
  /err:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Error.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Errorlog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Error
      tags:
      - Cube-Error
...
  /mold/bootstrap/plan:
    post:
      consumes:
      - application/json
      description: Mold 인프라 자동화 입력값을 검증하고 실행 순서, create command, 검증용 list command,
        누락 입력값을 반환합니다. hosts, primary/secondary storage, VLAN은 cluster.json의 type/hosts/ccvm.ip에서
        자동 계산하며 실제 인프라는 생성하지 않습니다.
      parameters:
      - description: Mold bootstrap request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.BootstrapRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
      summary: Mold bootstrap plan
      tags:
      - Mold
  /mold/capabilities:
    get:
      description: cluster.json의 clusterConfig.ccvm.ip로 Mold 기본 인증값(admin/password,
        domain=/)으로 login 후 listCapabilities command를 호출합니다. 인증값 override가 필요하면 POST
        /mold/capabilities를 사용합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
      summary: Mold capabilities 조회
      tags:
      - Mold
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: cluster.json의 clusterConfig.ccvm.ip로 Mold login 후 listCapabilities
        command를 호출합니다. 요청 body/form이 비어 있으면 admin/password, domain=/, response=json
        기본값을 사용합니다.
      parameters:
      - description: Mold login override
        in: body
        name: body
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.MoldAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
      summary: Mold capabilities 조회
      tags:
      - Mold
...
  /mold/session:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: cluster.json의 clusterConfig.ccvm.ip로 Mold endpoint를 계산한 뒤 login
        command를 호출해 sessionkey를 반환합니다. 요청 body/form이 비어 있으면 admin/password, domain=/,
        response=json 기본값을 사용합니다.
      parameters:
      - description: Mold login override
        in: body
        name: body
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.MoldAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_mold.Response'
      summary: Mold session 생성
      tags:
      - Mold
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"